int
Use int instead of size_t in MobiDoc and PalmDbReader

PdbReader dataSize, GetRecordCount, and GetRecord use int. MobiDoc internal decompression and record parsing follow suit, dropping redundant casts at call sites.

prompt: update MobiDoc.[h|cpp] and PalmDbReader.[h|cpp]; update callers to simplify int <=> size_t casting

int

+++ src/Menu.cpp
@@ -1160,7 +1160,7 @@ static void AddFileMenuItem(HMENU menuFile, Str filePath, int index) {
 
     TempStr menuString = path::GetBaseNameTemp(filePath);
     // shorten very long file names so that menu isn't too wide
-    const size_t kMaxRunes = 70;
+    const int kMaxRunes = 70;
     menuString = ShortenStringUtf8InTheMiddleTemp(menuString, kMaxRunes);
 
     TempStr fileName = MenuToSafeStringTemp(menuString);
